# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "News Portal API",
        "description": "API for News Portal",
        "contact": {
            "name": "Your Name",
            "email": "your@email.com"
        },
        "license": {
            "name": "Your License",
            "url": "http://your-license-url.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/paginated": {
            "get": {
                "tags": [
                    "News"
                ],
                "summary": "Search for news articles",
                "description": "Searches for news articles based on the provided parameters.",
                "operationId": "searchAllNews",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "The search query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sort option for the results",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "relevancy",
                                "popularity",
                                "publishedAt"
                            ]
                        }
                    },
                    {
                        "name": "language",
                        "in": "query",
                        "description": "Language code for filtering results",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "maxLength": 2
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "title": {
                                                "type": "string"
                                            },
                                            "description": {
                                                "type": "string"
                                            },
                                            "author": {
                                                "type": "string"
                                            },
                                            "publishedAt": {
                                                "type": "string"
                                            },
                                            "url": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/settings/show": {
            "get": {
                "tags": [
                    "User Settings"
                ],
                "summary": "Get user settings",
                "description": "Fetches the user's settings.",
                "operationId": "getUserSettings",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "defaultCategory": {
                                            "type": "string"
                                        },
                                        "articlesPerPage": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/settings/update": {
            "post": {
                "tags": [
                    "User Settings"
                ],
                "summary": "Update user settings",
                "description": "Updates the user's settings.",
                "operationId": "updateUserSettings",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "defaultCategory": {
                                        "type": "string",
                                        "enum": [
                                            "business",
                                            "entertainment",
                                            "general",
                                            "health",
                                            "science",
                                            "sports",
                                            "technology"
                                        ]
                                    },
                                    "articlesPerPage": {
                                        "type": "integer",
                                        "minimum": 5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Settings updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/paginated": {
            "get": {
                "tags": [
                    "Top News"
                ],
                "summary": "Get top news",
                "description": "Fetches top news based on the specified category and country.",
                "operationId": "getTopNews",
                "parameters": [
                    {
                        "name": "category",
                        "in": "query",
                        "description": "The category of news (business, entertainment, health, science, sports, technology).",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "business",
                                "entertainment",
                                "health",
                                "science",
                                "sports",
                                "technology"
                            ]
                        }
                    },
                    {
                        "name": "country",
                        "in": "query",
                        "description": "The country code for news filtering (e.g., 'us', 'gb', 'ca').",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "maxLength": 2
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "title": {
                                                "type": "string"
                                            },
                                            "description": {
                                                "type": "string"
                                            },
                                            "author": {
                                                "type": "string"
                                            },
                                            "publishedAt": {
                                                "type": "string"
                                            },
                                            "url": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}